fixes advanced search pane display glitch
Constraining rule
"fieldset .panel-body" not to apply to collapsible items by using 
"fieldset .panel-body:(.collapse)".
Problematic is the display:inherit, which should be display:none for collapsible items (toggles to display:block). Removing the display property is a bad choice either, as there might be flex items somewhere else on the page.